requirements.txt
langchain~=0.1.4
crewai~=0.1.32
python-dotenv~=1.0.1
pyyaml~=6.0.1
faiss-cpu~=1.7.4
pypdf~=4.0.2